Added typecast to "if (minBoundY == (uint8_t) ~0) return;" instances see: https://github.com/squix78/esp8266-oled-ssd1306/issues/140

+++ SH1106Brzo.h
@@ -83,7 +83,7 @@ class SH1106Brzo : public OLEDDisplay {
        // If the minBoundY wasn't updated
        // we can savely assume that buffer_back[pos] == buffer[pos]
        // holdes true for all values of pos
-       if (minBoundY == ~0) return;
+       if (minBoundY == (uint8_t) ~0) return;
 
        byte k = 0;
        uint8_t sendBuffer[17];
